\name{SPCALDA-package}
\alias{SPCALDA-package}
%\alias{SPCALDA}
\docType{package}
\title{
SPCALDA: A New Reduced-Rank Linear Discriminant Analysis Method
}
\description{
A new reduced-rank LDA method which works for high dimensional multi-class data.
}
\details{
\tabular{ll}{
Package: \tab SPCALDA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-04\cr
License: \tab GPL-2 \cr
}
}

\author{
Yue S. Niu, Ning Hao and Bin Dong.

Maintainer: Ning Hao <nhao@math.arizona.edu>
}

 
\keyword{ package }
 
